requirements.txt
python-dotenv
streamlit
langchain
google-generativeai